Move chartstatus resource from v2 to v3
