+++ service/controller/v2/resource_set.go
@@ -4,7 +4,6 @@ import (
 	"context"
 
 	"github.com/cenkalti/backoff"
-	"github.com/giantswarm/apiextensions/pkg/clientset/versioned"
 	"github.com/giantswarm/apprclient"
 	"github.com/giantswarm/helmclient"
 	"github.com/giantswarm/microerror"
@@ -17,7 +16,6 @@ import (
 
 	"github.com/giantswarm/chart-operator/service/controller/v2/key"
 	"github.com/giantswarm/chart-operator/service/controller/v2/resource/chart"
-	"github.com/giantswarm/chart-operator/service/controller/v2/resource/chartstatus"
 )
 
 const (
@@ -32,7 +30,6 @@ type ResourceSetConfig struct {
 	// Dependencies.
 	ApprClient apprclient.Interface
 	Fs         afero.Fs
-	G8sClient  versioned.Interface
 	HelmClient helmclient.Interface
 	K8sClient  kubernetes.Interface
 	Logger     micrologger.Logger
@@ -83,23 +80,8 @@ func NewResourceSet(config ResourceSetConfig) (*controller.ResourceSet, error) {
 		}
 	}
 
-	var chartStatusResource controller.Resource
-	{
-		c := chartstatus.Config{
-			G8sClient:  config.G8sClient,
-			HelmClient: config.HelmClient,
-			Logger:     config.Logger,
-		}
-
-		chartStatusResource, err = chartstatus.New(c)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-	}
-
 	resources := []controller.Resource{
 		chartResource,
-		chartStatusResource,
 	}
 
 	{
