+++ service/controller/v3/resource/chartstatus/create.go
@@ -7,7 +7,7 @@ import (
 	"github.com/giantswarm/helmclient"
 	"github.com/giantswarm/microerror"
 
-	"github.com/giantswarm/chart-operator/service/controller/v2/key"
+	"github.com/giantswarm/chart-operator/service/controller/v3/key"
 )
 
 func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
